requirements.txt
flask==2.0.1
requests==2.26.0
chatterbot==1.0.4
chatterbot-corpus==1.2.0
sentence-transformers
faiss-cpu
numpy
tqdm